requirements_install.txt
attrs==19.3.0
cffi==1.14.0
cryptography==2.8
importlib-metadata==1.5.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyaml==20.3.1
pycparser==2.20
pyparsing==2.4.6
pytest==5.4.1
pytest-assume==2.2.1
pytest-html==2.1.1
pytest-metadata==1.10.0
PyYAML==5.3.1
six==1.14.0
testinfra==5.0.0
wcwidth==0.1.8
zipp==3.1.0
